feat: 관심 종목 삭제에 접근 권한별 기능 추가

+++ src/main/java/com/stockleague/backend/stock/service/WatchlistService.java
@@ -93,6 +93,9 @@ public class WatchlistService {
         Watchlist watchlist = watchlistRepository.findById(watchlistId)
                 .orElseThrow(() -> new GlobalException(GlobalErrorCode.WATCHLIST_NOT_FOUND));
 
+        if (!watchlist.getUser().getId().equals(user.getId())) {
+            throw new GlobalException(GlobalErrorCode.FORBIDDEN);
+        }
 
         watchlistRepository.delete(watchlist);
